feat: solved task modules
